adapt csv parser
