+++ dojo/tools/openvas/xml_parser.py
@@ -1,3 +1,12 @@
+import csv
+import hashlib
+import io
+from dateutil.parser import parse
+from xml.dom import NamespaceErr
+from defusedxml import ElementTree as ET
+from dojo.models import Finding, Endpoint
+
+
 class OpenVASXMLParser(object):
     def __init__(self) -> None:
         pass
